requirements.txt
vanna
boto3
awscli
botocore
langchain
python-dotenv
